move tests to new package and fix import issues
